requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ── Data validation ────────────────────────────────────────────────────────────
# chromadb 0.4.x requires pydantic >=1.9, <3 — pin 2.6.x for stability
pydantic==2.6.4
pydantic-settings==2.2.1

# ── RAG / LangChain ────────────────────────────────────────────────────────────
langchain==0.2.1
langchain-community==0.2.1

# ── Embeddings ─────────────────────────────────────────────────────────────────
sentence-transformers==2.7.0
# torch CPU-only — install separately if you need GPU (see README)
torch==2.2.2

# ── Vector database ────────────────────────────────────────────────────────────
# 0.4.24 is the last stable release before the 0.5 API rewrite;
# it works cleanly with pydantic 2.x and the flatbuffers shipped by pip
chromadb==0.4.24

# flatbuffers pin — fixes the missing entry_points.txt error on Windows
flatbuffers==24.3.25

# ── DOCX parsing ──────────────────────────────────────────────────────────────
python-docx==1.1.2

# ── LLM provider (Groq) ───────────────────────────────────────────────────────
groq==0.9.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
httpx==0.27.0